# trace: use weak in TracerProvider for storing tracers