unique name
```
goos: darwin
goarch: arm64
pkg: go.opentelemetry.io/otel/sdk/trace
cpu: Apple M3
                       │ BenchmarkTracerWithUniqueName_output.txt │     afterUniqueName_output.txt      │
                       │                  sec/op                  │   sec/op     vs base                │
TracerWithUniqueName-8                                389.9n ± 9%   918.1n ± 8%  +135.44% (p=0.002 n=6)

                       │ BenchmarkTracerWithUniqueName_output.txt │ afterUniqueName_output.txt  │
                       │                   B/op                   │    B/op     vs base         │
TracerWithUniqueName-8                                543.5 ± 15%   550.0 ± 0%  ~ (p=1.000 n=6)

                       │ BenchmarkTracerWithUniqueName_output.txt │     afterUniqueName_output.txt      │
                       │                allocs/op                 │  allocs/op   vs base                │
TracerWithUniqueName-8                                 4.000 ± 0%   11.000 ± 0%  +175.00% (p=0.002 n=6)
```

same name:

```
goos: darwin
goarch: arm64
pkg: go.opentelemetry.io/otel/sdk/trace
cpu: Apple M3
                     │ BenchmarkTracerWithSameName_output.txt │      afterSameName_output.txt       │
                     │                 sec/op                 │   sec/op     vs base                │
TracerWithSameName-8                              20.84n ± 1%   51.05n ± 1%  +144.88% (p=0.002 n=6)

                     │ BenchmarkTracerWithSameName_output.txt │   afterSameName_output.txt    │
                     │                  B/op                  │    B/op     vs base           │
TracerWithSameName-8                               0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=6) ¹
¹ all samples are equal

                     │ BenchmarkTracerWithSameName_output.txt │   afterSameName_output.txt    │
                     │               allocs/op                │ allocs/op   vs base           │
TracerWithSameName-8                               0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=6) ¹
¹ all samples are equal
```