@yumosx 2025-04-13
HI @pellared, I used `weak`  to store the tracer (following the official Go blog's approach), but after performance testing, the results weren't good - performance actually decreased. 

Am I using the wrong approach here?