@dmathieu 2025-04-15
The linked issue in the description seems unrelated to this PR.

Note that we won't be able to merge this until 1.25 is released (which is when we will drop support for 1.23).