@pellared 2025-04-17
Hi, currently I have no time to review this PR, but I just want to call out that use of `weak` would not "improve" the performance (on the hot-path), but only give the garbage collector the possibility to remove tracers that are not referenced by the user.