@yumosx 2025-04-17
> use of `weak` would not "improve" the performance (on the hot-path). It will only give the garbage collector the possibility to remove tracers that are not referenced by the user. 

Yes, this is right, The [initial  Benchmark test](https://github.com/open-telemetry/opentelemetry-go/pull/6655#issue-2991192939) show `map[instrument.Scope]*tracer`  performs better than `weak.Pointer[tracer]`.

> So the total used memory should be lower (after garbage collection) when tracers are "used and forgotten" by the user (caller).

the second test,  forcing `runtime.GC()` shows the  weak+cleanup  version runs faster than without it.

I think this improvement comes from weak+cleanup lowering GC pressure.

finally, thank you for the insightful reply.